pyproject.toml
[tool.poetry]
name = "portal-universidade-api"
version = "0.1.0"
description = ""
authors = ["Luciano Gomes <lucianogvda02@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
djangorestframework = "^3.14.0"
black = "^23.9.1"
djangorestframework-simplejwt = "^5.3.0"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
